Time: 7 ms (1.81%) | Memory: 25.8 MB (15.78%) - LeetSync

+++ 3471-minimum-average-of-smallest-and-largest-elements/minimum-average-of-smallest-and-largest-elements.cpp
@@ -5,7 +5,7 @@ public:
         vector<float> averages;
         sort(nums.begin(), nums.end());
         for(int i = 0; i < n/2; i++){
-            float val = ((float)nums[i] + nums[n-i-1]) / 2;
+            float val = (nums[i] + nums[n-i-1]) / 2.0;
             averages.push_back(val);
         }
         float mini = INT_MAX;
